requirements.txt
fastapi
uvicorn
websockets
google-cloud-speech
pip install python-dotenv
